feat: new operator "^" added in StringCalc, now any calc with Calculator interface can be used for calculations in StringCalc, main in StringCalc replaced by not static method

+++ homework/src/homework3/Calculator.java
@@ -38,7 +38,7 @@ public interface Calculator {
      * @param factorTwo второй множитель
      * @return произведение множителей
      */
-    double multiplication(double factorOne, int factorTwo);
+    double multiplication(double factorOne, double factorTwo);
     
     /**
      * Нахождение разности двух чисел.
