+++ homework/src/homework3/CalculatorWithOperator.java
@@ -42,7 +42,7 @@ public class CalculatorWithOperator implements Calculator {
      * @param factorTwo второй множитель
      * @return произведение множителей
      */
-    public double multiplication(double factorOne, int factorTwo) {
+    public double multiplication(double factorOne, double factorTwo) {
         return factorOne * factorTwo;
     }
